Add ability to use full QuantumGraph in jobs.
In particular, this meant having options that indicated to not
create all the small job QuantumGraph files and the ability to
put node numbers on the pipetask command line.  With these,
also changed the implementation of the ClusteredQuantumGraph
to reduce memory usage.  Also made sure bpsUsedShared was
accurately reflected in the GenericWorkflow so HTCondor plugin
could use the large QuantumGraph in place instead of copying
it to every job.