Make pickling graphs optional
The graph with coalesced quanta and the generic workflow graph were
always serialized and saved for future reference using Python
pickle.dump(). However, it turned out that pickling large graphs
essentially doubles the memory footprint for a run submission.  As a
quick way to mitigate the problem, I introduced a flag which allow one
to disable pickling if necessary.