Determine the memory limit only if necessary
The memory autoscaling mechanism requires a hard memory limit to avoid
jobs staying in the job queue indefinitely due to unsatisfied memory
requirements.  BPS was trying to determine this limit by querying
available computational resources regardless whether the memory
autoscaling was enabled or not and always complained if failed.  I made
changes to prevent it from pestering users about it when not really
necessary.