Fix single concurrency limit splitting.
When exactly one concurrency_limit value, the bug was converting
the value into a comma-separated list of characters.  There was an
inconsistency across the code about whether the value was guaranteed
to be a list or not.  Changed all code to treat it as string, because
that is what was currently actually stored.