Try to keep naming convention consistent
Names of some local variables varied somewhat strongly between functions
though these variables were associated with objects of the same type.
While there was nothing wrong with the code because of it, I changed
names of some variables to keep them consistent and hopefully make the
code easier to follow.