Allow for specifying WMS plugin in environment
Commands like prepare, submit, restart, report, and cancel need to know
what WMS plugin to use.  This plugin could be specified either as a
command line argument or in the user config file (if not specified the
default value was used).  I added made a changes to allow to specify the
plugin using the environmental variable BPS_WMS_SERVICE.

The priorities are as follow:

1. Specifying the plugin at the command line (using ``--wms`` option)
   overrides any other specifications.
2. Specifying the plugin in the config file (using ``wmsServiceClass``)
   will take precedence over the value specified by the BPS_WMS_SERVICE
   (if any).
3. Specifying the plugin using BPS_WMS_SERVICE overrides the package
   default setting.