Set GenericWorfkloJob attributes consistently
Some job attributes of the GenericWokflowJob were only set for the job
representing the merging of the execution Butler. but not for "regular"
jobs executing a quantum or a quantum cluster.  Made changes to ensure
that the same subset of attributes will be correctly initialized for
each job in the GenericWorfklow.