Add a new job state, PRUNED
PanDA is capable of distinguishing between jobs that are not ready
(can't be run yet because their parents have not run/completed yet) and
jobs that won't run at all (at least one of their predecessors already
failed).  I added a new job state so this distinction can be captured
and properly reported by any WMS with such a capability.