Make using 'defaults' less confusing
To preserve the backward compatibility, using ``defaults`` optional
argument when instantiating a BpsConfig could be confusing, e.g., when
``None`` the configuration was being prepopulated with BPS and WMS
defaults, but only if the configuration was being read from a file).
Made the changes so, if provided, they will be included regardless of
how a BpsConfig is created.