Do not always include sections from search order
After DM-44110 the sections related to BpsConfig's search order are
always being included in BpsConfig by default. However, BpsConfig is
a hierarchical data structure that can include other configs as
values. As a result, these sections were being included in these
nested configs when they were returned as a search result.  Made
a change to prevent that and restore the old behavior of the
BpsConfig.search().