Make BPS config validation more flexible
Validating a BPS configuration for a regular workflow and for a
workflow executing a script remotely requires different sets of
validation tests.  However, the validation tests for BPS config
performed in init_submission() were "hard-coded" and there was no
mechanism allowing for customizing them.  Modified this function so
the validation tests it performs could be explicitly specified when
called.