Defer displaying report messages
Some of the messages coming from the reporting mechanism were displayed
as soon as they are available. As a result some of them were showing up
at very top of the output produced by ``bps report`` and could be easily
missed.  Made changes so all messages are printed at the very end of the
report.